+++ lib/deimos/schema_backends/avro_schema_registry.rb
@@ -21,14 +21,10 @@ module Deimos
 
       # @return [AvroTurf::Messaging]
       def avro_turf_messaging
-        path = Deimos.config.schema.path.presence || Deimos.config.schema.paths[:avro].first
-        if path.blank?
-          raise "No schema paths configured for `avro` backend!"
-        end
          @avro_turf_messaging ||= AvroTurf::Messaging.new(
           schema_store: @schema_store,
           registry_url: Deimos.config.schema.registry_url,
-          schemas_path: path,
+          schemas_path: Deimos.config.schema.path,
           user: Deimos.config.schema.user,
           password: Deimos.config.schema.password,
           namespace: @namespace,